+++ spec/main-spec.js
@@ -12,8 +12,47 @@
 
 'use strict'
 
-var customizeEngineHandlebars = require('../')
+var customize = require('customize')
+var _ = require('lodash')
 
-describe('customize-engine-handlebars:', function () {
-  // body
+var hb = customize()
+  .registerEngine('handlebars', require('../'))
+  .merge({
+    handlebars: {
+      partials: 'spec/fixtures/testPartials1',
+      helpers: {
+        helper1: function (abc) {
+          return 'helper1(' + abc + ')'
+        }
+      },
+      templates: 'spec/fixtures/templates',
+      data: {
+        eins: 'one',
+        zwei: 'two',
+        drei: 'three',
+        vier: 'four'
+      },
+      preprocessor: function (data) {
+        return _.mapValues(data, function (value) {
+          return '->' + value + '<-'
+        })
+      }
+    }
+  })
+
+describe('customize-engine-handlebars', function () {
+  it('should load partials and templates and return one result per template', function (next) {
+    hb.run().tap(function (result) {
+      expect(result).toEqual({
+        handlebars: {
+          'a.md': 'a.md testPartials1/eins one',
+          'b.md': 'b.md testPartials1/zwei two helper1([object Object])'
+        }
+      });
+
+    }).done(next)
+    // body
+  });
+
+
 })