+++ CHANGELOG.md
@@ -4,3 +4,5 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Upcoming
 ### Initial version
+
+- Extracted handlebars-engine from `customize`